+++ src/tracks/track_object_manager.cpp
@@ -193,12 +193,14 @@ void TrackObjectManager::enableFog(bool enable)
 // ----------------------------------------------------------------------------
 
 PhysicalObject* TrackObjectManager::insertObject(const std::string& model,
-                                                 PhysicalObject::bodyTypes shape, int mass, int radius,
+                                                 PhysicalObject::bodyTypes shape,
+                                                 float mass, float radius,
                                                  const core::vector3df& hpr,
                                                  const core::vector3df& pos,
                                                  const core::vector3df& scale)
 {
-    PhysicalObject* object = new PhysicalObject(model, shape, mass, radius, hpr, pos, scale);
+    PhysicalObject* object = new PhysicalObject(model, shape, mass, radius, 
+                                                hpr, pos, scale);
     object->init();
     m_all_objects.push_back(object);
     return object;
